Document and improve code comments across multiple classes
Added detailed Javadoc comments to several classes and methods, ensuring clarity on usage, parameters, and functionality. Minor formatting and consistency improvements were also made to enhance readability and maintainability.

+++ src/main/java/com/jmathanim/Utils/Vec.java
@@ -154,7 +154,7 @@ public class Vec implements Stateable, HasDirection {
     }
 
     public double norm() {
-        return (double) sqrt(x * x + y * y + z * z);
+        return sqrt(x * x + y * y + z * z);
     }
 
     /**
